+++ CHANGELOG.md
@@ -44,6 +44,8 @@ subcommand.
 - Upgraded all builds to use Go 1.10.
 - Use megacheck instead of errcheck.
 - Cleaned agent configuration.
+- We no longer duplicate hook execution for types that fall into both an exit
+code and severity (ex. 0, non-zero).
 
 ### Fixed
 - Fixed a bug in time.InWindow that in some cases would cause subdued checks to
@@ -55,6 +57,8 @@ erroneously expire.
 - Resolved a bug in how an executor processes checks. If a check contains proxy
 requests, the check should not duplicately execute after the proxy requests.
 - Removed an erroneous validation statement in check handler.
+- Fixed HookList `hooks` validation and updated `type` validation message to
+allow "0" as a valid type.
 
 ## [2.0.0-alpha.17] - 2018-02-13
 ### Added